improvements

+++ src/features/team-channels.js
@@ -34,12 +34,10 @@ export default {
                     let currentChannel = allChannels[i]
                     let previousChannel = allChannels[i-1]
 
-                    log.debug(currentChannel)
-
-                    log.debug("current: " + currentChannel.name + ":" + currentChannel.members.size)
-                    log.debug("previous: " + previousChannel.name + ":" + previousChannel.members.size)
-
-                    if(previousChannel){
+                    if(currentChannel && previousChannel){
+                        log.debug("current: " + currentChannel.name + ":" + currentChannel.members.size)
+                        log.debug("previous: " + previousChannel.name + ":" + previousChannel.members.size)
+                        
                         if(currentChannel.members.size == 0 && previousChannel.members.size == 0){
                             log.info("Deleting channel " + channel.name)
                             currentChannel.delete()
